initial upload
This is example code of my middleware services tester. The tester's web page calls this set of services, which in turn call the services set up by the developers for a phone app. When using this interface, the tester get POST/GET with a web page form and see the returned JSON data from the middleware.